# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Report Card Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: List scores
      parameters:
        - name: termId
          in: query
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: assessmentComponentId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
    put:
      operationId: upsertScore
      tags: [scores]
      summary: Upsert score (insert or update)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
  /scores/bulk:
    put:
      operationId: bulkUpsertScores
      tags: [scores]
      summary: Bulk upsert scores for a class/subject/term
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkScoreInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"

  # ── GRADING SCALES ────────────────────────────────────────────────────────
  /grading-scales:
    get:
      operationId: listGradingScales
      tags: [grading-scales]
      summary: List grading scales
      responses:
        "200":
          description: List of grading scale entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradingScale"
    post:
      operationId: createGradingScale
      tags: [grading-scales]
      summary: Create grading scale entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradingScaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradingScale"
  /grading-scales/{id}:
    patch:
      operationId: updateGradingScale
      tags: [grading-scales]
      summary: Update grading scale entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradingScaleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradingScale"
    delete:
      operationId: deleteGradingScale
      tags: [grading-scales]
      summary: Delete grading scale entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── REPORT CARDS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Score:
      type: object
      required: [id, studentId, assessmentComponentId, scoreValue]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        assessmentComponentId:
          type: integer
        componentName:
          type: string
        scoreValue:
          type: number
        isLocked:
          type: boolean
        enteredAt:
          type: string
        lastEditedAt:
          type: ["string", "null"]
    ScoreInput:
      type: object
      required: [studentId, assessmentComponentId, scoreValue]
      properties:
        studentId:
          type: integer
        assessmentComponentId:
          type: integer
        scoreValue:
          type: number
    BulkScoreInput:
      type: object
      required: [scores]
      properties:
        scores:
          type: array
          items:
            $ref: "#/components/schemas/ScoreInput"

    # ── Grading Scale ─────────────────────────────────────────────────────
    GradingScale:
      type: object
      required: [id, minScore, maxScore, gradeLabel, remark]
      properties:
        id:
          type: integer
        minScore:
          type: number
        maxScore:
          type: number
        gradeLabel:
          type: string
        remark:
          type: string
    GradingScaleInput:
      type: object
      required: [minScore, maxScore, gradeLabel, remark]
      properties:
        minScore:
          type: number
        maxScore:
          type: number
        gradeLabel:
          type: string
        remark:
          type: string
    GradingScaleUpdate:
      type: object
      properties:
        minScore:
          type: number
        maxScore:
          type: number
        gradeLabel:
          type: string
        remark:
          type: string

    # ── Report Cards ──────────────────────────────────────────────────────
...